requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.2
passlib==1.7.4
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
